# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: King SMM Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all SMM services
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user orders with optional filter
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersPage"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Insufficient balance or invalid input
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
...
  /wallet/topup:
    post:
      operationId: topupWallet
      tags: [wallet]
      summary: Request a wallet top-up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupInput"
      responses:
        "201":
          description: Top-up request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopupRequest"
...
components:
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        pricePerThousand:
          type: number
        minQuantity:
          type: integer
        maxQuantity:
          type: integer
        platform:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - name
        - category
        - description
        - pricePerThousand
        - minQuantity
        - maxQuantity
        - platform
        - isActive
        - platform
    Order:
      type: object
      properties:
        id:
          type: integer
        serviceId:
          type: integer
        serviceName:
          type: string
        platform:
          type: string
        link:
          type: string
        quantity:
          type: integer
        charge:
          type: number
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serviceId
        - serviceName
        - platform
        - link
        - quantity
        - charge
        - status
        - createdAt
    OrdersPage:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - totalPages
    OrderInput:
      type: object
      properties:
        serviceId:
          type: integer
        link:
          type: string
        quantity:
          type: integer
      required:
        - serviceId
        - link
        - quantity
...
    TopupInput:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        transactionId:
          type: string
      required:
        - amount
        - paymentMethod
        - transactionId
    TopupRequest:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        paymentMethod:
          type: string
        transactionId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - amount
        - paymentMethod
        - transactionId
        - status
        - createdAt
...